Dev/package name (#2)
* Change canfd message package name

* Change canfd package name

* Change package name

* Update execution name
