--- can_msgs/msg/Frame.msg
@@ -1,8 +1,0 @@
-Header header
-uint32 id
-bool is_rtr
-bool is_extended
-bool is_error
-uint8 dlc
-uint8[8] data
-
