+++ canfd_msgs/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">.
-  <name>can_msgs</name>
+  <name>canfd_msgs</name>
   <version>0.8.5</version>
   <description>CAN related message types.</description>
 
@@ -8,10 +8,6 @@
   <author email="ivor@iwanders.net">Ivor Wanders</author>
 
   <license>BSD</license>
-
-  <url type="website">http://wiki.ros.org/can_msgs</url>
-  <url type="repository">https://github.com/ros-industrial/ros_canopen</url>
-  <url type="bugtracker">https://github.com/ros-industrial/ros_canopen/issues</url>
 
   <buildtool_depend>catkin</buildtool_depend>
