+++ ros_canopen/package.xml
@@ -16,6 +16,7 @@
   <buildtool_depend>catkin</buildtool_depend>
 
   <exec_depend>can_msgs</exec_depend>
+  <exec_depend>canfd_msgs</exec_depend>
   <exec_depend>canopen_402</exec_depend>
   <exec_depend>canopen_chain_node</exec_depend>
   <exec_depend>canopen_master</exec_depend>
