+++ socketcanfd_bridge/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>socketcan_bridge</name>
+  <name>socketcanfd_bridge</name>
   <version>0.8.5</version>
   <description>Conversion nodes for messages from SocketCAN to a ROS Topic and vice versa.</description>
 
@@ -16,6 +16,7 @@
   <buildtool_depend>catkin</buildtool_depend>
 
   <depend>can_msgs</depend>
+  <depend>canfd_msgs</depend>
   <depend>rosconsole_bridge</depend>
   <depend>roscpp</depend>
   <depend>socketcan_interface</depend>
