+++ socketcanfd_bridge/src/rosconsole_bridge.cpp
