+++ socketcanfd_bridge/test/initialize_vcan.sh
